refactor(none): ✨ arrange under the components directory
