+++ src/components/common/Footer/Footer.astro
@@ -1,6 +1,6 @@
 ---
 import { getEntry } from 'astro:content';
-import { Icon } from '@/components/Icon';
+import { Icon } from '@/components/ui/Icon';
 import { ACCOUNT_ID, GITHUB_BASE_URL, UNSPLASH_BASE_URL } from '@/lib/consts';
 import { getLocaleFromUrl, useTranslatedPath } from '@/utils/i18n/utils';
