+++ src/components/common/Navigation/Navigation.astro
@@ -1,9 +1,9 @@
 ---
 import '@/styles/pixelmplus.css';
 import { getEntry } from 'astro:content';
-import { Icon } from '@/components/Icon';
+import { Icon } from '@/components/ui/Icon';
 import { LocalePicker } from '@/components/LocalePicker/LocalePicker';
-import { Modal } from '@/components/Modal/Modal';
+import { Modal } from '@/components/ui/Modal/Modal';
 import { Search } from '@/components/Search/Search';
 import { transparentButton } from '@/styles/styles.css';
 import { getLocaleFromUrl, useTranslatedPath } from '@/utils/i18n/utils';
