+++ src/components/elements/Callout/CalloutTitle.astro
@@ -1,5 +1,5 @@
 ---
-import { Icon } from '@/components/Icon';
+import { Icon } from '@/components/ui/Icon';
 import { generateUniqueId } from '@/utils/generate-unique-id';
 import { toBoolean } from '@/utils/to-boolean';
 import type { Props as CalloutProps } from './Callout.astro';
