+++ src/components/models/BlogList/BlogList.astro
@@ -1,7 +1,7 @@
 ---
 import type { CollectionEntry } from 'astro:content';
+import { Icon } from '@/components/ui/Icon';
 import { FormattedDate } from '@/components/models/FormattedDate';
-import { Icon } from '@/components/Icon';
 import { Taxonomy } from '@/components/models/Taxonomy';
 import { getBlogCategory, getBlogTags } from '@/lib/collections/data';
 import { getLocaleFromUrl, useTranslatedPath } from '@/utils/i18n/utils';
