+++ src/components/models/BuyMeACoffee/BuyMeACoffee.astro
@@ -1,7 +1,7 @@
 ---
 import { getEntry } from 'astro:content';
-import { Icon } from '@/components/Icon';
-import { Modal } from '@/components/Modal/Modal';
+import { Icon } from '@/components/ui/Icon';
+import { Modal } from '@/components/ui/Modal/Modal';
 import { ACCOUNT_ID } from '@/lib/consts';
 import { getLocaleFromUrl } from '@/utils/i18n/utils';
 import {
