+++ src/components/models/Pager/Pager.astro
@@ -3,7 +3,7 @@ import { type CollectionEntry, getEntry } from 'astro:content';
 import { DEFAULT_PAGINATION_WIDTH } from '@/lib/consts';
 import { getLocaleFromUrl } from '@/utils/i18n/utils';
 import type { Page } from 'astro';
-import { Icon } from '../Icon';
+import { Icon } from '@/components/ui/Icon';
 
 type Props = {
   page?: Page<CollectionEntry<'blog' | 'news'>>;
