+++ src/components/models/PhotoGallery/PhotoGalleryItem.astro
@@ -1,12 +1,12 @@
 ---
 import { getEntry } from 'astro:content';
+import { ProfileIcon } from '@/components/ui/ProfileIcon';
 import { Card } from '@/components/elements/Card';
-import { ProfileIcon } from '@/components/ProfileIcon';
 import type { UnsplashPhoto } from '@/lib/unsplash/types';
 import { getTimeDiff } from '@/utils/get-time-diff';
 import { getLocaleFromUrl } from '@/utils/i18n/utils';
 import { toSnakeCase } from '@/utils/to-snake-case';
-import { Image } from '@unpic/astro';
+import { Image } from 'astro:assets';
 
 type Props = {
   photo: UnsplashPhoto;
