+++ src/components/models/Toc/TocDrawer.astro
@@ -1,6 +1,6 @@
 ---
 import { getEntry } from 'astro:content';
-import { Icon } from '@/components/Icon';
+import { Icon } from '@/components/ui/Icon';
 import { getLocaleFromUrl } from '@/utils/i18n/utils';
 
 const locale = getLocaleFromUrl(Astro.url);
