+++ src/components/ui/Balloon/Balloon.astro
@@ -1,7 +1,7 @@
 ---
 import { getEntry } from 'astro:content';
 import profileImage from '@/assets/images/profile.png';
-import { ProfileIcon } from '@/components/ProfileIcon';
+import { ProfileIcon } from '@/components/ui/ProfileIcon';
 import { getLocaleFromUrl } from '@/utils/i18n/utils';
 import type { ImageMetadata } from 'astro';
