+++ src/components/ui/ProfileIcon/ProfileIcon.astro
@@ -1,5 +1,5 @@
 ---
-import { Image } from '@unpic/astro';
+import { Image } from 'astro:assets';
 import type { ImageMetadata } from 'astro';
 import type { HTMLAttributes } from 'astro/types';
