+++ src/layouts/NewsLayout.astro
@@ -1,9 +1,9 @@
 ---
 import type { CollectionEntry } from 'astro:content';
-import { Article } from '@/components/Article';
+import { Article } from '@/components/models/Article';
+import { Icon } from '@/components/ui/Icon';
 import { BuyMeACoffee } from '@/components/models/BuyMeACoffee';
 import { FormattedDate } from '@/components/models/FormattedDate';
-import { Icon } from '@/components/Icon';
 import BaseLayout from '@/layouts/BaseLayout.astro';
 import { getSlugWithoutLocale } from '@/utils/get-slug-without-locale';
 import { getLocaleFromUrl, useTranslatedPath } from '@/utils/i18n/utils';
