+++ src/layouts/PageLayout.astro
@@ -1,8 +1,8 @@
 ---
 import type { CollectionEntry } from 'astro:content';
-import { Article } from '@/components/Article';
+import { Article } from '@/components/models/Article';
+import { Icon } from '@/components/ui/Icon';
 import { FormattedDate } from '@/components/models/FormattedDate';
-import { Icon } from '@/components/Icon';
 import BaseLayout from '@/layouts/BaseLayout.astro';
 import { getLocaleFromUrl } from '@/utils/i18n/utils';
