+++ src/lib/mdx-components.ts
@@ -1,11 +1,11 @@
-import { Balloon } from '@/components/Balloon';
-import { StarRating } from '@/components/StarRating';
 import { Callout, CalloutTitle } from '@/components/elements/Callout';
 import { Card, CardGrid } from '@/components/elements/Card';
 import { Figure } from '@/components/elements/Figure';
 import { Link } from '@/components/elements/Link';
 import { OEmbed } from '@/components/elements/OEmbed';
 import { FormattedDate } from '@/components/models/FormattedDate';
+import { Balloon } from '@/components/ui/Balloon';
+import { StarRating } from '@/components/ui/StarRating';
 
 export const mdxComponents = {
   a: Link,
