+++ src/pages/[locale]/blog/tags/[tag]/[...page].astro
@@ -1,7 +1,7 @@
 ---
 import { getEntry } from 'astro:content';
-import { BlogList } from '@/components/BlogList';
-import { Pager } from '@/components/Pager';
+import { BlogList } from '@/components/models/BlogList';
+import { Pager } from '@/components/models/Pager';
 import BaseLayout from '@/layouts/BaseLayout.astro';
 import {
   getLocaleContentEntries,
