+++ src/pages/[locale]/index.astro
@@ -1,9 +1,9 @@
 ---
 import { getEntry } from 'astro:content';
-import { BlogList } from '@/components/BlogList';
+import { BlogList } from '@/components/models/BlogList';
+import { Icon } from '@/components/ui/Icon';
+import { PhotoGallery } from '@/components/models/PhotoGallery';
 import { BulletinBoard } from '@/components/models/BulletinBoard';
-import { Icon } from '@/components/Icon';
-import { PhotoGallery } from '@/components/PhotoGallery';
 import BaseLayout from '@/layouts/BaseLayout.astro';
 import { getSortedContentEntries } from '@/lib/collections/contents';
 import {
